bbf: add support anywan type upstream iface in igmp
It was observed on dg400prime that the uci config was not correctly
generated when wan was of type anywan, this has been taken care of in
this patch. For anywan type IP interface, the list of upstream interfaces
is generated to keep each interface in mind. So, Device.IP.Interface.2
in this case will generate
	list upstream_interface 'eth5.1'
	list upstream_interface 'atm0.1'
	list upstream_interface 'ptm0.1'

test:
- Verify that config is correctly generated for normal interface.
- Verify that config is correctly generated for anywan interface.
- Verify that config is correctly generated for bridge interface.